+++ example/vite.client.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
     outDir: resolve(__dirname, "dist/_nix-js"),
     emptyOutDir: true,
     rollupOptions: {
-      input: resolve(__dirname, "src/entry-client.ts"),
+      input: resolve(__dirname, ".nix-js/entry-client.ts"),
       output: {
         entryFileNames: "entry-client.js",
         format: "es",